FIX KIRBY! SCROLL_Y was the issue. Need to make sure we're correctly using the render line versus the sprite line Y offset

+++ src/gameboy/gfx/tile_render_options.rs
@@ -8,11 +8,11 @@ pub struct TileRenderOptions {
     pub render_type: TileRenderType,
     pub map_addr: usize,
     pub tile_base_addr: usize,
-    pub line: usize
+    pub line: u8
 }
 
 impl TileRenderOptions {
-    pub fn new(render_type: TileRenderType, line: usize, map_addr: usize, tile_base_addr: usize) -> TileRenderOptions {
+    pub fn new(render_type: TileRenderType, line: u8, map_addr: usize, tile_base_addr: usize) -> TileRenderOptions {
         TileRenderOptions {
             render_type: render_type,
             line: line,
